backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.51.0
pypdf>=5.0.0
python-multipart>=0.0.9
